+++ asteria/src/instantiated_function.hpp
@@ -15,11 +15,11 @@ class Instantiated_function : public Abstract_function
   private:
     Vector<String> m_params;
     String m_file;
-    Uint64 m_line;
+    Uint32 m_line;
     Block m_body;
 
   public:
-    Instantiated_function(Vector<String> params, String file, Uint64 line, Block body) noexcept;
+    Instantiated_function(Vector<String> params, String file, Uint32 line, Block body) noexcept;
     ~Instantiated_function();
 
   public:
